modifs fronts et back, récupère des points pondérés mais pas arrondis

+++ app/models/basketballmatch.rb
@@ -57,8 +57,8 @@ class Basketballmatch < ApplicationRecord
       match_odds = call_api(end_point)["response"][0]["bookmakers"][0]["bets"][0]["values"]
       if match_odds[1]["odd"].present?
         game.update(
-          points_home: get_odd(match_odds, "Home").to_i * 10,
-          points_away: get_odd(match_odds, "Away").to_i * 10
+          points_home: get_odd(match_odds, "Home").to_f * 10,
+          points_away: get_odd(match_odds, "Away").to_f * 10
         )
       end
     end
