+++ app/models/sport_odd.rb
@@ -55,9 +55,9 @@ class SportOdd
     end_point = URI("#{BASE_URL}odds/fixture/#{game.fixture_id}")
     match_odds = call_api(end_point)["api"]["odds"][0]["bookmakers"][1]["bets"][0]["values"]
     game.update(
-      points_home: get_odd(match_odds, "Home").to_i * 10,
-      points_draw: get_odd(match_odds, "Draw").to_i * 10,
-      points_away: get_odd(match_odds, "Away").to_i * 10
+      points_home: get_odd(match_odds, "Home").to_f * 10,
+      points_draw: get_odd(match_odds, "Draw").to_f * 10,
+      points_away: get_odd(match_odds, "Away").to_f * 10
     )
   end
